to fix time update

+++ app/src/main/java/com/example/android/playlistmaker/player/domain/api/AudioPlayerInteractor.kt
@@ -13,6 +13,6 @@ interface AudioPlayerInteractor {
     }
 
     interface AudioPlayerTrackTimeConsumer {
-        fun getTime(time: Long)
+        fun getTime(time: Int)
     }
 }
